implement handlebar helpers

+++ controllers/api/comment-routes.js
@@ -1,26 +1,8 @@
 const router = require('express').Router();
-const { Comment, User, Post } = require('../../models');
+const { Comment } = require('../../models');
 
 router.get('/', (req, res) => {
-console.log('======================');
-  Comment.findAll({
-    attributes: [
-        'id',
-        'comment_text',
-        'created_at',
-      ],
-      order: [['created_at', 'DESC']],
-      include: [
-        {
-            model: User,
-            attributes: ['username']
-          },
-        {
-          model: Post,
-          attributes: ['title']
-        }
-      ]
-    })
+  Comment.findAll()
     .then(dbCommentData => res.json(dbCommentData))
     .catch(err => {
       console.log(err);
@@ -30,35 +12,39 @@ console.log('======================');
 
 router.post('/', (req, res) => {
   // expects => {comment_text: "This is the comment", user_id: 1, post_id: 2}
-  Comment.create({
-    comment_text: req.body.comment_text,
-    user_id: req.body.user_id,
-    post_id: req.body.post_id
-  })
-    .then(dbCommentData => res.json(dbCommentData))
-    .catch(err => {
-      console.log(err);
-      res.status(400).json(err);
-    });
+  if (req.session) {
+    Comment.create({
+      comment_text: req.body.comment_text,
+      user_id: req.session.user_id,
+      post_id: req.body.post_id
+    })
+      .then(dbCommentData => res.json(dbCommentData))
+      .catch(err => {
+        console.log(err);
+        res.status(400).json(err);
+      });
+  }
 });
 
 router.delete('/:id', (req, res) => {
-  Comment.destroy({
-    where: {
-      id: req.params.id
-    }
-  })
-    .then(dbCommentData => {
-      if (!dbCommentData) {
-        res.status(404).json({ message: 'No comment found with this id!' });
-        return;
+  if (req.session) {
+    Comment.destroy({
+      where: {
+        id: req.params.id
       }
-      res.json(dbCommentData);
     })
-    .catch(err => {
-      console.log(err);
-      res.status(500).json(err);
-    });
+      .then(dbCommentData => {
+        if (!dbCommentData) {
+          res.status(404).json({ message: 'No comment found with this id!' });
+          return;
+        }
+        res.json(dbCommentData);
+      })
+      .catch(err => {
+        console.log(err);
+        res.status(500).json(err);
+      });
+  }
 });
 
 module.exports = router;
